put auth functions back in NavBar
